Ctrl-z now deselects UI from event system so that textbox editing end function can trigger. Anchor bpm history tracking now working on moving and deleting. Still need to do clipboard pasting and decided whether to allow anchors to be moved/copied.

+++ Moonscraper Chart Editor/Assets/Scripts/ActionHistory.cs
@@ -22,15 +22,18 @@ public class ActionHistory
 
     public void Insert(Action[] action)
     {
-        // Clear all actions above the history point
-        actionList.RemoveRange(historyPoint + 1, actionList.Count - (historyPoint + 1));
-        timestamps.RemoveRange(historyPoint + 1, timestamps.Count - (historyPoint + 1));
+        if (action.Length > 0)
+        {
+            // Clear all actions above the history point
+            actionList.RemoveRange(historyPoint + 1, actionList.Count - (historyPoint + 1));
+            timestamps.RemoveRange(historyPoint + 1, timestamps.Count - (historyPoint + 1));
 
-        // Add the action in
-        actionList.Add(action);
-        timestamps.Add(Time.time);
+            // Add the action in
+            actionList.Add(action);
+            timestamps.Add(Time.time);
 
-        ++historyPoint;
+            ++historyPoint;
+        }
     }
 
     public void Insert(Action action)
